p2-Enhanced-Desktop/requirements.txt
# P2 - Enhanced Desktop Requirements
# Enhanced version with YOLOv8, orientation correction, and intelligent caching

# Core dependencies from P1
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
easyocr>=1.7.0
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0  # For Hough transform
Pillow>=10.0.0
requests>=2.31.0
streamlit>=1.28.0

# YOLOv8 Object Detection
ultralytics>=8.0.0

# Caching System
redis>=4.5.0
# sqlite3  # Built-in Python

# Performance Monitoring
psutil>=5.9.0
# cProfile  # Built-in Python

# Async Processing
# asyncio  # Built-in Python

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Optional: GPU support for YOLOv8
# torch with CUDA support (Linux/Windows)
# torch>=2.0.0 --index-url https://download.pytorch.org/whl/cu118